forgery_detection_service/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
pillow
pytesseract
scipy
